+++ freshbooks/__init__.py
@@ -10,12 +10,7 @@ The FreshBooks Python SDK allows you to more easily utilize the [FreshBooks API]
 from freshbooks.builders.filter import FilterBuilder  # noqa
 from freshbooks.builders.includes import IncludesBuilder  # noqa
 from freshbooks.builders.paginator import PaginateBuilder  # noqa
+from freshbooks.builders.sort import SortBuilder  # noqa
 from freshbooks.client import Client  # noqa
 from freshbooks.errors import FreshBooksError  # noqa
 from freshbooks.models import VisState  # noqa
-
-__pdoc__ = {
-    "freshbooks.api.resource": False,
-    "freshbooks.builders.paginator.PaginateBuilder.MIN_PAGE": False,
-    "freshbooks.builders.paginator.PaginateBuilder.MAX_PER_PAGE": False
-}