+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added upload attachment and image resources
 - Fixed `invoice_payment_options` create call (was not creating)
 - Updated webhook event error handling for new FreshBooks API error strcuture
+- Added list sort builder
 
 ## 1.0.1