+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.1.0
+
 - Added upload attachment and image resources
 - Fixed `invoice_payment_options` create call (was not creating)
 - Updated webhook event error handling for new FreshBooks API error strcuture